admin-web/requirements.txt
alembic==0.9.1
appdirs==1.4.3
bleach==2.0.0
blinker==1.4
click==6.7
Flask==0.12
Flask-Breadcrumbs==0.4.0
Flask-Login==0.4.0
Flask-Menu==0.5.1
flask-paginate==0.4.5
Flask-Principal==0.4.0
Flask-WTF==0.14.2
html5lib==0.999999999
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
PyMySQL==0.7.10
pyparsing==2.2.0
python-dateutil==2.6.0
python-editor==1.0.3
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.6
webencodings==0.5
Werkzeug==0.12
WTForms==2.1